OpenAI-Chatbot/requirements.txt
langchain-openai
langchain
langchain_community
python-dotenv 
langchain_core
streamlit